Added readme file and fix editlist

+++ database/migrations/2022_01_07_053055_create_data_table.php
@@ -15,7 +15,7 @@ class CreateDataTable extends Migration
     {
         Schema::create('data', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('value');
+            $table->integer('value');
             $table->date('date');
             $table->timestamps();
         });
